Change integer distribution signature, add extensions.

+++ src/main/scala/ru/ifmo/onell/util/BinomialDistribution.scala
@@ -5,9 +5,9 @@ import java.util.Random
 import ru.ifmo.onell.IntegerDistribution
 
 object BinomialDistribution {
-  def apply(n: Int, p: Double): IntegerDistribution = rng => withScanner(n, p, rng)
+  def apply(n: Long, p: Double): IntegerDistribution = rng => withScanner(n, p, rng)
 
-  private def withScanner(n: Int, p: Double, rng: Random): Int = {
+  private def withScanner(n: Long, p: Double, rng: Random): Int = {
     val sc = BinomialScanner(p)
     var idx = sc.offset(rng) - 1
     var result = 0
